manifests: drop coreos-assembler-claim2 PVC
We haven't used it in a while and don't really miss it. Since it's
easier to not deal with it in the new cluster, let's just formalize its
removal.

This doesn't purge support for it in the primary pipeline job yet. Let's
do that separately from the Fedora migration underway.

Drop `simple-httpd` also since its only purpose was browing the cache.