Jenkinsfile.release: pass --build to cosa meta
We hit this issue where one pipeline run kicked off immediately after
another one had finished. So we had something like:

- next-devel-x86_64#1
- next-devel-aarch64#1
- next-devel-x86_64-release#1
- next-devel-x86_64#2
- next-devel-aarch64-release#1

At the point that the release job runs for `#1` after the aarch64
pipeline is done the builds.json has already been updated with build
ID from the `#2` run. In this case `cosa meta` will look at the meta.json
from the wrong build.

To work around this we need to just be explicit and pass the build
ID directly to `cosa meta`.