jobs/build: also early archive QEMU image
It's already in `meta.json` so we can't *not* upload it. Otherwise if
there's a build failure before the full upload, it can cause issues.
E.g. a broken link in the build browser and throwing off upgrade tests.
We could remove it from the JSON but that seems unnecessary.