jobs: only run release job once per build
Before, we were running the release job multiple times per build: once
for each architecture. This was done by triggering it at the end of both
the x86_64 build and multi-arch builds.

However, it's possible that the x86_64 build fails while the multi-arch
build succeeds. In that case, we don't want to release just for e.g.
aarch64 if the x86_64 build is missing or broken.

Fix this by making only the x86_64 build trigger the release job. This
way, it'll only be triggered once and only if the x86_64 build
succeeded. Add locking in the release job and the multi-arch job to
ensure that the release job cannot execute until the multi-arch job has
completed, whether successfully or not.

This also allows us to hack around issues in the tools used by the
release job which don't fully support releasing for a single
architecture only. These are gaps that we should still fix regardless.