build/release: allow passing arches from toplevel build job
This will allow us to specify ADDITIONAL_ARCHES to the toplevel build
job and have it filter all the way through the cascading secondary jobs.
It also means that we're using a single source of truth (the parameter)
for each job rather than mixing the paramater and the additional_arches
definition from `streams.groovy`.