Drop logic for developer_prefix in pipelines
The whole concept is a bit confusing and there is a decent amount of
conditional logic in there. If someone needs to do some hacking for
development we'll just give them their own project in openshift and
a fresh s3 bucket in another account. That way there's no risk of
messing with the official releases, but they are exercising similar
code paths.

officialfix