manifests/jenkins: enable OVERRIDE_PV_PLUGINS_WITH_IMAGE_PLUGINS
Jenkins stores configs and plugins in a single directory, which is on a
PVC mount. When updating the Jenkins image, by default new plugins in
the image are not copied over into the PVC. If we want to make the image
plugins authoritative (i.e. if we actually want plugin updates), we need
to opt into it via an env var.