jobs/build: cut image building parallelism in half
We're seeing issues where the pipeline sometimes hits EAGAIN on `fork`:

https://github.com/coreos/fedora-coreos-tracker/issues/1186

This may have started happening more frequently after the recent
addition of VirtualBox.

Anyway, it's kinda silly to build that many images in parallel. There's
diminishing returns and eventually a performance cost. So let's do it in
two rounds instead. (Ideally instead it'd be a kind of task pool, but
let's not try to implement that in Groovy.)