jobs/build: unparallelize keylime hash generation
There's an issue with `flufl.Lock` where it fails when exiting the
locked context with `NotLockedError: Already unlocked` when something
else was trying to lock it.

We'll keep digging into that, but for now let's just unparallelize these
to get releases going again. It doesn't save us much time anyway.