jobs/build: pass arch explicitly; don't use AWS_CONFIG_FILE env var
The is prep for us making our build and build-arch jobs more similar.
Here we start passing the architecture around in more places like we
are doing in build-arch and also stop using the $AWS_CONFIG_FILE env
var in favor of passing it over the CLI. Passing via CLI means that
it will survive being proxied to a remote session.