jobs/build: insert parent info via `cosa meta`
Rather than update the JSON file in groovy directly here
in the pipeline let's update it in the `cosa` context by
leveraging the already existing `cosa meta` command.

This will enable us to have parity with what we'll do in
build-arch and the `cosa remote-session` work will allow
`cosa meta` to transparently work on the remote builder.