jobs/build: use cosa shell for command execution
Filtering our commands through `cosa shell` means that we can
use the exact same code for our main pipeline as we will need
to use for our build-arch job.

We also modify checkKolaSuccess() here to just take a full file
path to the report and copy the report back individually to
report.json in the top level dir.