jobs/bump-lockfile: use new --find-build-for-arch buildfetch option
Right now the latest build in a stream could have failed the secondary
architectures runs and have missing arches. This causes the `cosa build`
to fail because it tries to read the meta data for the latest build
but none exists for this arch. Let's have buildfetch traverse the
build history to find the latest build and fetch that build's info
for each respective architecture.