jobs/build: Set Force=true for forked multi-arch builds
We've seen it happen a few times recently where the x86_64 build
detects a change and does a build but the multi-arch builds don't
detect a change and report "No apparent changes since previous commit".

This is probably because of our COSA builds being out of sync, which
we'll fix soon, but either way we probably want to set Force=true
for our multi-arch builds.