manifests/jenkins.yaml: set OVERRIDE_PV_CONFIG_WITH_IMAGE_CONFIG
By default, when using a PVC, the OpenShift S2I Jenkins image only
instantiates templates on the first run. This however means that
configuration state can accumulate in the PVC.

Set the `OVERRIDE_PV_CONFIG_WITH_IMAGE_CONFIG` env var which tells the
S2I run script to always instantiate and copy in templates. This
*should* preserve build logs.

An example of this is the Kubernetes cloud config, which was lost in a
recent change (see #583).