jobs/build-cosa: limit to a single simultaneous run for now
The builders aren't super beefy and are shared for pipeline builds
so let's disallow multiple concurrent cosa builds at the same time.

We can relax this in the future if we like, but for now there appears
to be a scenario when starting jenkins from scratch that we could try
to build for all supported branches at the same time and we need to
hedge against 10 builds getting submitted to the builders at the same
time.