jobs/build: request as many CPUs as parallel VMs
Right now we're running with `--parallel 5` and run the upgrade tests in
parallel, but only have 4 CPUs. This causes some heavier tests to fail.
Let's request as many CPUs as we run VM tests in parallel. Since the
latter is calculated from the memory request, that's what we do.

In effect, this bumps up our CPU requests to 6.