jobs/build-cosa: push in v2s2 format
The cosa image we build is used to trigger internal RHCOS multi-arch
rebuilds (and in the near future will be used directly once we have all
the arches built). But the stack on the target cluster is too old to
understand OCI, so we get:

    Internal error occurred: missing signature key

Add `--v2s2` to push in v2s2 format.

Once we've rebased RHCOS to the FCOS pipeline, we should be able to
revert this.

Related: https://github.com/coreos/coreos-assembler/pull/2726