Simplify root CA cert provisioning
With the Jenkins creds approach, the Kubernetes secret name needs to
match what we expect, so we can simplify things. For example, instead of
having to pass `--additional-root-ca-cert-secret` to `deploy`, just have
the latter detect it automatically. We can also simplify the
`jenkins-agent` template further.

That way, the only thing one has to do is create the Kubernetes secret,
and everything else will Just Work.