tree: convert `streams.groovy` to `config.yaml`
We want to move to a model where the pipeline is configured using a sane
configuration format. The backbone of this config format will be the
streams listing, which currently is embedded in `streams.groovy`.

Delete `streams.groovy`. Create `config.yaml`, maintaining the same
information that was in the former. Adapt the rest of the codebase as
necessary.