Rename `pipecfg` to `jenkinscfg`
I'd like to use the term "pipeline config" and the shorthand `pipecfg`
to refer to the higher-level configuration bits which dictate the actual
e.g. streams to build, etc...

Rename the current "pipeline config" to "jenkins config". I.e. this is
used to confgure the lower-level Jenkins-y bits of the pipeline. Any
higher-level config should live in `config.yaml`.

Many knobs currently in the pipeline configmap shouldn't be there. They
will be moved to `config.yaml` in a future patch.