Support pipecfg living in a separate repo
As part of https://github.com/coreos/fedora-coreos-pipeline/issues/618,
we need to move to a model where the pipeline config is stored outside
of this repo.

Add support for this via a new `./deploy --pipecfg` option which takes a
git repo containing a `config.yaml` in the root of the repo.