tree: move FCOS pipecfg to separate repo
This repo currently hosts both the FCOS pipeline, and the FCOS pipeline
configuration. We need to untangle these two as we make this pipeline
more generic in order to support building RHCOS on top of it.

Move `config.yaml` to a new repo. This new repo is temporarily hosted
at:

https://github.com/jlebon/fedora-coreos-pipecfg

Move the `next-devel` machinery along with it.

Drop support for reading the pipeline config in-tree.

For more information, see:
https://github.com/coreos/fedora-coreos-pipeline/issues/618