jobs/build-fcos-buildroot: use Generic Webhook Trigger plugin
This plugin allows us to only trigger the job when something in
the ci/buildroot/ subdirectory has changed, which will cut down
on wasted resources. We also add a cron build trigger here to
make sure the job runs at least a few times a week to keep the
packages fresh in the buildroot.