jobs/build-arch: bridge `cosa remote-session` to older RHCOS releases
We recently did some golang migration [1] and happened to add the
`cosa remote-session` pieces [2] after that migration. This makes
it not easy to backport to other branches. For a period of time
let's use the latest COSA for our COSA pod for multi-arch builds
instead of the versioned COSA. We will still use the versioned COSA
on the remote so the actual building will happened with the correctly
versioned COSA (substituted below and passed as an argument to
`cosa remote-session create --image=`).

[1] https://github.com/coreos/coreos-assembler/pull/2919
[2] https://github.com/coreos/coreos-assembler/pull/2979