config.yaml: drop `type: fcos` knob
Let's not have code that explicitly switches on FCOS vs RHCOS in the
pipeline. Instead, let's try to take each instance on a case-by-case
basis and formalize the distinction they're trying to express at a
higher level.

In the case of the aarch64 serial console issue, we'll get that via the
recent `platforms.yaml` work. For now, to not regress let's introduce a
temporary knob to activate the AWS aarch64 serial console hack.