jobs/build-arch: start cosa remote session earlier
I.e. start it before we `git ls-remote` for the source commit. This
makes no practical difference for now because those commands don't run
on the remote. But it's prep for injecting the root CA in both the
controller and remote cosa environments, and thus needs to be done
before `git ls-remote`.