convert remaining jobs to use cosaPod
Let's switch the rest of our jobs to use cosaPod() to gain consistency.
The last remaining reason to keep the old model (secrets mounting in
pod.yaml) no longer remains because we are no longer mounting secrets
in our cosaPod but rather binding using jenkins credentials.

This also means we can delete pod.yaml, because it is no longer used.