jobs: fix `cosaPod` scope and reorder
In fa556b2 ("convert remaining jobs to use cosaPod"), we changed scopes
so that the cosa container no longer covered the `finally` block. But we
emit a fedmsg there with the final result of the job, so we do still
need the cosa container. Fix this by re-ordering the scopes again.