docs: document config.yaml schema
We're adding a lot of keys to `config.yaml` right now. We need to make
sure we do a proper job of documenting them so that it's easier for
others to make use of the file.

Let's document the keys we currently support and from now on require any
new keys to be documented there.