jobs/{build,kola*}: bump memory requests
Now that we set the memory limit as well we find out we're using
more than we are requesting. Let's request a more appropriate amount
of memory for our jobs.

For the build job we set it to 8.5 Gi because we run two testiso runs
at the same time and those can take up to 4Gi each.

For the kola-* jobs the request is updated to 1Gi if the job is going
to download and decompress the disk image artifact and 512Mi otherwise.