uploads: update powervs config and document
"cloud_object_storage" isn't very descriptive. I originally didn't
quite understand what it was referring to, but after some research
I understand it to be referring to the cloud object storage service
instance that the buckets are associated with. Let's make the config
entry more descriptive at this level.

Also let's add documentation for how "bucket" is treated in the context
of powervs.