Document image push secret for old registry location
For FCOS, we're temporarily pushing the oscontainer to
`quay.io/coreos-assembler/fcos`. RHCOS will have the same setup, where
we'll need to push the oscontainer to `registry.ci.openshift.org` for
4.11 and earlier. 4.11 isn't going EOL anytime soon, so let's document
the feature properly, which includes documenting the secret. And
accordingly move the FCOS-specific note from the pipeline code to the
config file.