jobs/{build,build-arch}: only run tests if artifacts exist
Let's conditionalize running ISO tests and cloud tests if the
necessary artifacts for those tests exist in our build metadata
for the current run.

Structuring it like this enables us to easily comment out artifacts
in the pipeline config.yaml and seamlessly disable running the tests
for the artifact rather than having to comment out code.