config: add cloud.test_architectures key
This key is used to limit which architecures we attempt to run cloud
tests for on a given cloud/platform. For example, we create OpenStack
artifacts for every architecture, but we don't have an OpenStack that
we can test in that has all of those architectures so we need to limit
the architectures that those tests are attempted on.