jobs/build-arch: aws_aarch64_serial_console_hack check sooner
With the current logic a pipeline needs to exclude `aws` in the list
of artifacts it includes in the config.yaml even if it doesn't need
the hack. Let's re-work the logic so pipelines that don't need the
hack don't need to exclude `aws` in the aarch64 artifact list. This
also means those pipelines won't need to be updated when the hack
is removed.