libupload: don't buildextend with --force for azure
If we make it to the cloud uploads in the pipeline then we've
reserved a build ID and any new runs will reserved a different
build ID. We shouldn't need to pass `--force` here.

In it's current form we're building the azure image in the
"Build Artifacts" stage and again in the "Cloud Upload" stage
befause `--force` tells `buildextend-azure` to ignore the
existing built artifact.