Drop support for coreos-assembler ImageStream
We want to move the coreos-assembler image to use to the pipecfg, and
having to declare an image upfront at pipeline initialization time
clashes with this.

This does mean we're no longer caching the image in-cluster. In practice
however, we find that the Kubernetes is very efficient at reusing nodes
with the latest content so that the majority of the time, no pull is
necessary.

We update HACKING.md, and while we're here, also document `--pipecfg`.