Support configuring cosa image in pipecfg
Follow-up to the previous commit. We add support for configuring the
cosa image from the pipecfg. We support both a top-level knob and a
stream-level knob, which will be necessary for development RHCOS streams
not yet using a branched cosa.

Keep the `COREOS_ASSEMBLER_IMAGE` parameter available for on-the-fly
overrides when needed.