jobs/release: add Fetch Artifacts stage
For the Push Container stage we need to have all of those artifacts
actually fetched to the local pod. Because of the way buildfetch works
(stomping on meta.json) we need to run the fetch of the artifacts before
any stages that modify/update the meta.json file. Let's add a new stage
that will handle this for us before any of the later stages that modify
meta.json.