utils: enhance per-stream artifact set modifications
RHCOS 4.11 and older do not want to build the `extensions-container`
artifact. The artifact is for all arches. Support `all` for
`skip_artifacts` to make skipping it less verbose.

RHCOS 4.11 and older want to build the `legacy-oscontainer` artifact.
RHCOS 4.12 and newer do not. Support enhancing the default set by
bringing back the `additional_artifacts` key. The artifact is for all
arches. Support the `all` key to make adding it less verbose.

For consistency with the other keys, also support `additional_artifacts`
per architecture.