jobs/release: use remote builder for release job
We are bumping up into an issue where we can't run `podman` in
the kubernetes pod inside the cluster [1]. We use `podman` to create
the manifest lists and push containers to the registries.

To workaround this problem for now we'll just run the release job
from a remote builder where we have more permissions/control.

[1] https://github.com/coreos/fedora-coreos-pipeline/issues/723