Support pushing containers in v2s2 format
A bunch of tooling around OpenShift don't support the OCI format yet, so
for the RHCOS images we need to push in v2s2 format.

For FCOS, we want to keep pushing in OCI format. Add a new config.yaml
knob to override this for RHCOS.

Related: https://github.com/coreos/coreos-assembler/pull/2726