tree-wide: use different home dir to workaround podman mount issue
This is to address https://github.com/coreos/fedora-coreos-pipeline/issues/723

One workaround I've found is to change my `$HOME` to a different
filesystem (`/var/tmp`, which is overlayfs) and it seems to workaround
the offending check [1].

This shouldn't work because there is another check [2] that should fail
if we're on an overlay filesystem, but it doesn't in this case. I suspect
we're hitting some kind of corner case.

[1] https://github.com/containers/storage/blob/232bf398bd68595d2a6e53ec07d8b3c4424be3c0/drivers/overlay/check.go#L158-L164
[2] https://github.com/containers/storage/blob/232bf398bd68595d2a6e53ec07d8b3c4424be3c0/drivers/overlay/overlay.go#L314-L318