jobs/build: add EARLY_ARCH_JOBS parameter
This will give us control of when the multi-arch jobs are kicked off.
In some cases we want them to be kicked off early (to save time).
Unfortunately in that case if the x86_64 build later fails (which
means the build ID can no longer be used) the work done by the
multi-arch builds gets trashed as well.

Let's add a knob so we can kick them off early when desired, but
allow for them to be kicked off later in situations where time isn't
a large factor.