jobs/release: allow pushing both legacy and new oscontainer
The latest plan for 4.12 is to go back to shipping both the legacy and
new oscontainers to not break existing workflows of teams consuming it.
So we need to undo some of the recent changes made.

Related: https://issues.redhat.com/browse/MCO-295
Related: https://issues.redhat.com/browse/MCO-392