ci: add basic syntax checking of Groovy files
I realized recently that we could use `load()` in CI as a way to
syntax-check our Groovy files. Add a CI check that uses this.

This works for regular Jenkinsfiles too (since it's still just the same
custom Groovy interpreter). But `load()` will try to execute them, which
is not what we want. Hack around this by adding a `return this` at the
top. The parser will still actually scan the full file.

This only catches basic syntax errors like unbalanced parentheses or
brackets, and not semantic errors like typoed function names or
variables. We should be able to add a more complete CI run here in the
future that e.g. could execute a part of the pipeline, but it requires
some thought.