jobs/kola-*: drop `S3_STREAM_DIR` parameter
This parameter dates from the very start of kola jobs. At that time, we
didn't pass in the stream itself as a separate parameter, so we passed
the whole S3 stream dir instead.

Nowadays, we do pass in the stream, and all kola jobs correctly query
the pipecfg for that info. So clean things up by deleting the
unnecessary parameter.