jobs/release: pass `--build` parameter to `cosa meta`
In the release job, we may not be releasing the latest build but our use
of `cosa meta` assumed so. Pass `--build` to be more strict.

We currently only call `run_cloud_tests()` from contexts where it's
reasonable to expect the target version is the latest build, but as it's
structured as a standalone function, also pass `--build` there to be
safe.