Rework how `additional_arches` works
We need to make clearer the distinction between the set of architectures
that a Jenkins instance supports, and the set of architectures that a
pipecfg wants to build for. This is usually the same, but not always.
For example, a development stream may be temporarily unhealthy on some
arches, or a hotfix pipecfg may want to limit the set of arches to build
for.

Let's have the semantics of `additional_arches` be similar to that of
`cosa_img`: it's a config knob that can be overridden by a parameter.
Accordingly, set the job parameter default to be empty (matching
`COREOS_ASSEMBLER_IMAGE`).

Notice how this cleans up the `build-cosa` job, which no longer needs to
read in the pipecfg. This makes sense since it's not strictly part of
the build pipeline.

This also fixes a subtle issue with the way the arch-related params were
wired. Their defaults were derived from the pipecfg, but those default
values only actually affect the *next* run of the job. To be correct, we
should've been using the definition from the pipecfg throughout the job,
not the parameter.