Support skipping or overriding additional arches per stream
Add support for `skip_additional_arches` and `additional_arches` keys in
the stream-level section. E.g. currently for FCOS, we're not shipping
ppc64le for the production streams.

This will also be useful for RHCOS in the future when e.g. keeping older
streams alive on only a subset of arches.