jobs/release: don't push base-oscontainer on releases missing it
For RHCOS 4.11 and older, there is no native oscontainer to push so we
don't want to even try. But unlike other cloud-y things, the artifact
does exist (`ostree`), so we can't just rely on the fact that we're not
building it.

Rather than adding more temporary knobs at the pipecfg, hack around this
by checking if the cosa schema supports `base-oscontainer` and if not,
don't try to push it.