add hacks.use_yumrepos_branch_workaround
For the moment the git repo that we are pulling our yum repo files
from has a branch per stream of RHCOS (i.e. 4.9, 4.10...). The plan
is to get to a single branch that has all the yum repo files. Until
that happens let's bake in a hack to support what currently exists.