jobs: run pods with `jenkins` service account
Run the pods we provision with the same service account as the one
Jenkins itself runs with. This should probably be something done by
default by the kubernetes-plugin, but surprisingly AFAICT there are no
knobs for this.

Anyway, in the future we could look at "usptreaming" this into
coreos-ci-lib directly, though I need to look into what's the cleanest
way to introspect the service account being used in a generic way.

This fixes the fact that with the default SA, we can't actually query
annotations from the Slack secret for getting the emoji prefix to use.