jobs/build: support ADDITIONAL_ARCHES value 'none'
We need a way to express an override to turn off all additional arches.
Also modify `release` and `cloud-replicate` for consistency.

This also makes the latter use `pipeutils.get_additional_arches()`. I
wasn't sure if there was a reason it didn't use it like we do in the
other jobs.