config: drop support for `override_compression_format` hack
We don't need this anymore. On FCOS, the `xz` setting is part of
`image.yaml` which `cosa compress` now knows to read from. On RHCOS, we
rely on `cosa compress` defaulting to `gzip`. This works for both new
and old RHCOS streams using a cosa without the `image.yaml` support.