libcloud: check for `govcloud` config before replicating
After e101235, we support disabling AWS GovCloud uploads by dropping the
associated `clouds` config. We need to check for this in the replication
code too because the `meta.amis` check is not enough.

An alternative to this that would be more consistent with the pattern in
this function is checking for specific GovCloud regions in `meta.amis`,
though that code would be more complex.