manifests: rename `jenkins-agent` BuildConfig and template
The `jenkins-agent` BuildConfig outputs an image that is basically the
same as the `jenkins-agent-base` image from OpenShift but with a cert,
so use `jenkins-agent-base-with-cert` as its name to be clear.

While we're here, rename the template filename to `jenkins-with-cert` to
match (a future patch will add another buildconfig in there for the
controller too).