jobs/build-arch: use new autolock feature
This makes use of the new autolock feature in cosa to force multi-arch
builds to be consistent with the x86_64 build. This takes effect when
building a stream which doesn't currently use lockfiles.

The primary use case is RHCOS, but even once we start using lockfiles
there, this will still be useful for the FCOS rawhide stream.